third_party/blink/web_tests/platform/generic/fast/writing-mode/flipped-blocks-hit-test-box-edges-expected.txt

x	y	element	range start container
25	25	outer	after-after
25	100	after-after	after-after
49	25	outer	after
49	100	after	after
50	25	outer	middle
50	100	middle	middle
149	25	outer	middle
149	100	middle	middle
150	25	outer	before
150	100	before	before
165	25	outer	before-before
165	100	before-before	before-before
